functions/requirements.txt
# Azure Functions runtime
azure-functions==1.21.3

# Azure Identity & Authentication
azure-identity==1.19.0
msal==1.31.0

# Microsoft Graph SDK (beta - required for Intune APIs)
msgraph-beta-sdk>=1.12.0

# Log Analytics ingestion & querying
azure-monitor-ingestion==1.0.4
azure-monitor-query==1.4.0

# Note: Versions pinned for reproducibility (Jan 2026)
# Run `pip list --outdated` to check for updates
